add some error handling

+++ dotter/mainwindow.cpp
@@ -3,6 +3,8 @@
 #include <QSignalMapper>
 #include <QString>
 
+#include <cassert>
+
 #include "abortwidget.h"
 #include "app.h"
 #include "config.h"
@@ -12,7 +14,7 @@
 
 MainWindow::MainWindow(QWidget* parent) :
         QMainWindow(parent), _ui(new Ui_MainWindow), _svgView(new SvgView()), _abortWidget(
-                new AbortWidget()), _process(new QProcess(this)), _toolMapper(
+                new AbortWidget()), _toolMapper(
                 new QSignalMapper(this))
 {
     _ui->setupUi(this);
@@ -35,17 +37,14 @@ MainWindow::MainWindow(QWidget* parent) :
     setCentralWidget(container);
     _abortWidget->hide();
 
-    connect(_process, SIGNAL(finished(int, QProcess::ExitStatus)), this,
-            SLOT(layoutFinished(int, QProcess::ExitStatus)));
-
     connect(_toolMapper, SIGNAL(mapped(const QString&)), this,
             SLOT(setLayout(const QString&)));
 
-    _ui->actionLayoutDot->setProperty("layout", "dot");
+    _ui->actionLayoutDot->setProperty(  "layout", "dot");
     _ui->actionLayoutTwopi->setProperty("layout", "twopi");
     _ui->actionLayoutCirco->setProperty("layout", "circo");
     _ui->actionLayoutNeato->setProperty("layout", "neato");
-    _ui->actionLayoutFdp->setProperty("layout", "fdp");
+    _ui->actionLayoutFdp->setProperty(  "layout", "fdp");
 
     auto group = new QActionGroup(this);
     for (auto action :
@@ -64,6 +63,7 @@ MainWindow::MainWindow(QWidget* parent) :
 
 MainWindow::~MainWindow()
 {
+    stopLayout();
     delete _ui;
 }
 
@@ -97,8 +97,6 @@ void MainWindow::openFile(const QString& fileName)
 
 void MainWindow::layoutFinished(int exitCode, QProcess::ExitStatus exitStatus)
 {
-    _abortWidget->hide();
-
     switch (exitStatus)
     {
         case QProcess::NormalExit:
@@ -110,7 +108,7 @@ void MainWindow::layoutFinished(int exitCode, QProcess::ExitStatus exitStatus)
             else
             {
                 QString msg = QString("%1 returned with exit code %2.\n%3\n%4");
-                msg = msg.arg(_process->program());
+                msg = msg.arg(_lastCommandLine);
                 msg = msg.arg(exitCode);
                 msg = msg.arg(QString(_process->readAllStandardOutput()));
                 msg = msg.arg(QString(_process->readAllStandardError()));
@@ -119,21 +117,53 @@ void MainWindow::layoutFinished(int exitCode, QProcess::ExitStatus exitStatus)
             break;
         case QProcess::CrashExit:
             QString msg = QString("%1 has crashed.\n%2\n%3");
-            msg = msg.arg(_process->program());
+            msg = msg.arg(_lastCommandLine);
             msg = msg.arg(QString(_process->readAllStandardOutput()));
             msg = msg.arg(QString(_process->readAllStandardError()));
             QMessageBox::critical(this, "dotter", msg);
             break;
     }
+
+    stopLayout();
+}
+
+void MainWindow::layoutError(QProcess::ProcessError error)
+{
+    switch (error)
+    {
+        case QProcess::FailedToStart:
+        {
+            QString msg = QString("Failed to start %1.");
+            msg = msg.arg(_lastCommandLine);
+            QMessageBox::critical(this, "dotter", msg);
+            break;
+        }
+        default:
+        {
+            QString msg = QString("Unknown error when running %1.");
+            msg = msg.arg(_lastCommandLine);
+            QMessageBox::critical(this, "dotter", msg);
+            break;
+        }
+    }
+
+    stopLayout();
 }
 
 void MainWindow::stopLayout()
 {
     _abortWidget->hide();
-    if (_process->state() != QProcess::NotRunning)
+
+    if (_process)
     {
-        _process->kill();
-        _process->waitForFinished();
+        if (_process->state() != QProcess::NotRunning)
+        {
+            _process->kill();
+            _process->waitForFinished();
+        }
+
+        delete _process;
+        _process = nullptr;
     }
 }
 
@@ -141,21 +171,40 @@ void MainWindow::startLayout()
 {
     stopLayout();
 
-    if (!_fileName.isEmpty())
+    assert(!_process);
+
+    if (_fileName.isEmpty())
     {
-        Config* config = static_cast<App*>(qApp)->config();
-        const QString exePath = config->getToolPath(_tool);
-
-        _abortWidget->setMessage(
-                QString("Computing '%1' layout for '%2'...").arg(_tool,
-                        _fileName));
-        _abortWidget->show();
-        QStringList proc_args;
-        proc_args.push_back("-Tsvg");
-        proc_args.push_back(_fileName);
-
-        _process->start(exePath, proc_args);
+        return;
     }
+
+    _process = new QProcess(this);
+    connect(_process, SIGNAL(finished(int, QProcess::ExitStatus)), this,
+            SLOT(layoutFinished(int, QProcess::ExitStatus)));
+
+    connect(_process, &QProcess::errorOccurred, this, &MainWindow::layoutError);
+
+    Config* config = static_cast<App*>(qApp)->config();
+    const QString exePath = config->getToolPath(_tool);
+    if (exePath.isEmpty())
+    {
+        QString msg = QString("No configured executable path for tool '%1'.");
+        msg = msg.arg(_tool);
+        QMessageBox::critical(this, "dotter", msg);
+        return;
+    }
+
+    _abortWidget->setMessage(
+            QString("Computing '%1' layout for '%2'...").arg(_tool,
+                    _fileName));
+    _abortWidget->show();
+    QStringList proc_args;
+    proc_args.push_back("-Tsvg");
+    proc_args.push_back(_fileName);
+
+    _lastCommandLine = QString("%1 -tsvg %2").arg(exePath).arg(_fileName);
+
+    _process->start(exePath, proc_args);
 }
 
 void MainWindow::setLayout(const QString& layout)
